chore: split scpi utils into seperate file
